feat: added possibility to add global queries or share queries with roles

+++ routes/users.php
@@ -1012,12 +1012,31 @@ Route::post('/crud/users/approve', function () {
 
 Route::post('/crud/queries', function () {
     include_once BASEPATH . "/php/init.php";
-    if (isset($_POST['id'])) {
+    
+    $action = $_POST['action'] ?? 'ADD';
+
+    if (isset($_POST['id']) && $action == 'DELETE') {
         // delete query with _id
         $deleteResult = $osiris->queries->deleteOne(['_id' => DB::to_ObjectID($_POST['id'])]);
         return $deleteResult->getDeletedCount();
         die;
     }
+    if ($action == 'SHARE'){
+        if (isset($_POST['global'])) {
+            $values = ['global' => true];
+        } else if (isset($_POST['role'])) {
+            $values = ['global' => false, 'role' => $_POST['role']];
+        } else {
+            die("no visibility given");
+        }
+        $updateResult = $osiris->queries->updateOne(
+            ['_id' => DB::to_ObjectID($_POST['id'])],
+            ['$set' => $values]
+        );
+        return $updateResult->getModifiedCount();
+        die;
+    }
+
     if (!isset($_POST['name'])) die("no name given");
     if (!isset($_POST['rules'])) die("no rules given");
     if (!isset($_SESSION['username'])) die("no user given");
